+++ Sources/7zCoder.swift
@@ -52,20 +52,3 @@ class SevenZipCoder {
     }
 
 }
-
-extension SevenZipCoder: Equatable {
-
-    static func == (lhs: SevenZipCoder, rhs: SevenZipCoder) -> Bool {
-        let propertiesEqual: Bool
-        if lhs.properties == nil && rhs.properties == nil {
-            propertiesEqual = true
-        } else if lhs.properties != nil && rhs.properties != nil {
-            propertiesEqual = lhs.properties! == rhs.properties!
-        } else {
-            propertiesEqual = false
-        }
-        return lhs.id == rhs.id && lhs.numInStreams == rhs.numInStreams &&
-            lhs.numOutStreams == rhs.numOutStreams && propertiesEqual
-    }
-
-}